--- a/src/Sdcb.OpenVINO.Extensions.OpenCvSharp4/MatExtensions.cs
+++ b/src/Sdcb.OpenVINO.Extensions.OpenCvSharp4/MatExtensions.cs
@@ -79,12 +79,14 @@ public static class MatExtensions
     /// The resulted Mat is made by vertically stacking the input Mats, each having the same type and specified height and width.
     /// </summary>
     /// <param name="srcs">An array of Mats that are to be stacked. It's assumed that all provided Mats are having the same type.</param>
+    /// <param name="height">no need anymore,keep it just for compatibility</param>
+    /// <param name="width">no need anymore,keep it just for compatibility</param>
     /// <returns>Returns a new Mat that is a vertical stack of all input Mats, each having the specified height and width.</returns>
     /// <exception cref="ArgumentException">Thrown when srcs is null or empty, or when any of the Mats in srcs does not match the specified type.</exception>
     /// <remarks>
     /// This method utilizes OpenCvSharp4 for operations on Mats. Pay attention to match correctly the type of Mats in the 'srcs' array.
     /// </remarks>
-    public static Mat StackingVertically(this Mat[] srcs)
+    public static Mat StackingVertically(this Mat[] srcs, int height, int width)
     {
         Mat combinedMat = new();
         Cv2.VConcat(this.Mats, combinedMat);