hi @sdcb
sorry for that I didn't noticed they have variety size.

I thought they already have the same size after padding and resize.

I looked into your implementation of  **PaddleOcrRecognizer**.

https://github.com/sdcb/OpenVINO.NET/blob/6b99f896651c5b7336a8ee33a7670cfcb6703f21/projects/PaddleOCR/Sdcb.OpenVINO.PaddleOCR/PaddleOcrRecognizer.cs#L129-L144

the input of model must have the same shape in one batch.

I think you have no doubt about this.

It should be like the image below after invoked the method **PrepareAndStackImages**.

![stacking](https://github.com/sdcb/OpenVINO.NET/assets/5615543/08da6f37-9c2e-4147-82a7-e81953336c1d)

Am I right? Here's my question.

the implementation of **[PrepareAndStackImages](https://github.com/sdcb/OpenVINO.NET/blob/6b99f896651c5b7336a8ee33a7670cfcb6703f21/projects/PaddleOCR/Sdcb.OpenVINO.PaddleOCR/PaddleOcrRecognizer.cs#L113)** invoked the method **[ResizePadding](https://github.com/sdcb/OpenVINO.NET/blob/6b99f896651c5b7336a8ee33a7670cfcb6703f21/projects/PaddleOCR/Sdcb.OpenVINO.PaddleOCR/PaddleOcrRecognizer.cs#L241C25-L241C38)**

why not the method of **ResizePadding** resize and pad image to **maxWidth** directly?

then we can invoke **VConcat** immediately.

https://github.com/sdcb/OpenVINO.NET/blob/6b99f896651c5b7336a8ee33a7670cfcb6703f21/projects/PaddleOCR/Sdcb.OpenVINO.PaddleOCR/PaddleOcrRecognizer.cs#L208-L239

according to what you mentioned in  last PR (#5) , I made this new PR , 

It fixs all problem, you can take a look.

I confirmed It keeps **exactly the same** OCR result before this PR (same input and same model).

meanwhile we also can get some improvement as shown below.

```

BenchmarkDotNet v0.13.10, Windows 10 (10.0.19044.3693/21H2/November2021Update)
  [Host]     : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-PARHUX : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-YLHAAW : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

IterationCount=10  LaunchCount=1  WarmupCount=1  

```
| Method                   | Runtime  | modelHeight | maxWidth | Mean     | Error    | StdDev   | Ratio | RatioSD | Allocated | Alloc Ratio |
|------------------------- |--------- |------------ |--------- |---------:|---------:|---------:|------:|--------:|----------:|------------:|
| **StackingVerticallyBySdcb** | **.NET 7.0** | **48**          | **320**      | **415.2 μs** | **15.14 μs** | **10.01 μs** |  **1.00** |    **0.00** |   **1.93 KB** |        **1.00** |
| StackingVerticallyByAven | .NET 7.0 | 48          | 320      | 252.8 μs |  8.97 μs |  5.93 μs |  0.61 |    0.02 |   1.79 KB |        0.93 |
|                          |          |             |          |          |          |          |       |         |           |             |
| StackingVerticallyBySdcb | .NET 8.0 | 48          | 320      | 408.3 μs | 11.43 μs |  7.56 μs |  1.00 |    0.00 |   1.93 KB |        1.00 |
| StackingVerticallyByAven | .NET 8.0 | 48          | 320      | 248.5 μs | 12.37 μs |  8.18 μs |  0.61 |    0.02 |   1.79 KB |        0.93 |
|                          |          |             |          |          |          |          |       |         |           |             |
| **StackingVerticallyBySdcb** | **.NET 7.0** | **48**          | **512**      | **423.3 μs** | **18.53 μs** | **12.26 μs** |  **1.00** |    **0.00** |   **1.54 KB** |        **1.00** |
| StackingVerticallyByAven | .NET 7.0 | 48          | 512      | 361.4 μs | 14.72 μs |  9.74 μs |  0.85 |    0.04 |   1.62 KB |        1.05 |
|                          |          |             |          |          |          |          |       |         |           |             |
| StackingVerticallyBySdcb | .NET 8.0 | 48          | 512      | 419.8 μs | 14.95 μs |  9.89 μs |  1.00 |    0.00 |   1.54 KB |        1.00 |
| StackingVerticallyByAven | .NET 8.0 | 48          | 512      | 360.3 μs | 14.88 μs |  9.84 μs |  0.86 |    0.03 |   1.62 KB |        1.05 |