@sdcb 2023-11-28
I agreed that we can use the method of ResizePadding resize and pad image to maxWidth directly in previous code, but I believe the StackingVertically method should keeps the same behavior of not constraint the input `srcs` width, which allows input `srcs` able to be different width.