@sdcb 2023-11-28
@AvenSun Hi AvenSun, I created a branch and did some necessary work that I believe, can you merge my upstream https://github.com/sdcb/OpenVINO.NET/tree/mat-ext branch into your PR's branch? After that, I'll go ahead and merge the PR.