@AvenSun 2023-11-28
hi @sdcb 
I just did what you said, you can have a try.